requirements.txt
discord.py
motor
python-dotenv
requests
dnspython
coc.py
